GradeBook/BookTests.cs
using Xunit;

namespace GradeBook
{
    public class BookTests
    {
        [Fact]
        public void BookCalculatesAnAverageGrade()
        {
            // skapa
            var book = new Book("");
            book.AddGrade(89.1);
            book.AddGrade(90.5);
            book.AddGrade(77.3);

            // Agera
            var result = book.GetStatistics();

            // Assert (Assert är en test class som testar olika kondition för den givande resultat)
            Assert.Equal(85.6, result.Average, 1);
            Assert.Equal(90.5, result.High, 1);
            Assert.Equal(77.3, result.Low, 1);
        }
    }
}
